+++ modules/examples/src/main/java/com/ibm/cloud/platform_services/iam_policy_management/v1/IamPolicyManagementExamples.java
@@ -134,7 +134,7 @@ public class IamPolicyManagementExamples {
       Response<Policy> response = service.createPolicy(options).execute();
       Policy policy = response.getResult();
 
-      System.out.println(policy);
+      System.out.printf("createPolicy() result:%n%s%n", policy.toString());
 
       // end-create_policy
 
@@ -154,7 +154,7 @@ public class IamPolicyManagementExamples {
       Response<Policy> response = service.getPolicy(options).execute();
       Policy policy = response.getResult();
 
-      System.out.println(policy);
+      System.out.printf("getPolicy() result:%n%s%n", policy.toString());
 
       // end-get_policy
 
@@ -216,7 +216,7 @@ public class IamPolicyManagementExamples {
       Response<Policy> response = service.updatePolicy(options).execute();
       Policy policy = response.getResult();
 
-      System.out.println(policy);
+      System.out.printf("updatePolicy() result:%n%s%n", policy.toString());
 
       // end-update_policy
 
@@ -257,7 +257,7 @@ public class IamPolicyManagementExamples {
       Response<PolicyList> response = service.listPolicies(options).execute();
       PolicyList policyList = response.getResult();
 
-      System.out.println(policyList);
+      System.out.printf("listPolicies() result:%n%s%n", policyList.toString());
 
       // end-list_policies
     } catch (ServiceResponseException e) {
@@ -272,8 +272,8 @@ public class IamPolicyManagementExamples {
               .policyId(examplePolicyId)
               .build();
 
-      service.deletePolicy(options).execute();
-
+      Response<Void> response = service.deletePolicy(options).execute();
+      System.out.printf("deletePolicy() response status code: %d%n", response.getStatusCode());
       // end-delete_policy
     } catch (ServiceResponseException e) {
       logger.error(String.format("Service returned status code %s: %s\nError details: %s",
@@ -294,7 +294,7 @@ public class IamPolicyManagementExamples {
       Response<CustomRole> response = service.createRole(options).execute();
       CustomRole customRole = response.getResult();
 
-      System.out.println(customRole);
+      System.out.printf("createRole() result:%n%s%n", customRole.toString());
 
       // end-create_role
 
@@ -314,7 +314,7 @@ public class IamPolicyManagementExamples {
       Response<CustomRole> response = service.getRole(options).execute();
       CustomRole customRole = response.getResult();
 
-      System.out.println(customRole);
+      System.out.printf("getRole() result:%n%s%n", customRole.toString());
 
       // end-get_role
 
@@ -337,7 +337,7 @@ public class IamPolicyManagementExamples {
       Response<CustomRole> response = service.updateRole(options).execute();
       CustomRole customRole = response.getResult();
 
-      System.out.println(customRole);
+      System.out.printf("updateRole() result:%n%s%n", customRole.toString());
 
       // end-update_role
     } catch (ServiceResponseException e) {
@@ -355,7 +355,7 @@ public class IamPolicyManagementExamples {
       Response<RoleList> response = service.listRoles(options).execute();
       RoleList roleList = response.getResult();
 
-      System.out.println(roleList);
+      System.out.printf("listRoles() result:%n%s%n", roleList.toString());
 
       // end-list_roles
     } catch (ServiceResponseException e) {
@@ -370,7 +370,8 @@ public class IamPolicyManagementExamples {
         .roleId(exampleCustomRoleId)
         .build();
 
-      service.deleteRole(options).execute();
+      Response<Void> response = service.deleteRole(options).execute();
+      System.out.printf("deleteRole() response status code: %d%n", response.getStatusCode());
 
       // end-delete_role
     } catch (ServiceResponseException e) {